// src/common/enums/permission.type.ts
export const PermissionTypeDetails = {
// ...
  DEFESA: { designacao: "Defesa", sigla: "mgtfc_p_dfsa" },
// ...
} as const;